[glue] move DataTrainingArguments around
