+++ tests/test_trainer.py
@@ -1,6 +1,6 @@
 import unittest
 
-from transformers import AutoTokenizer, DataTrainingArguments, TrainingArguments, is_torch_available
+from transformers import AutoTokenizer, TrainingArguments, is_torch_available
 
 from .utils import require_torch
 
@@ -14,6 +14,7 @@ if is_torch_available():
         DefaultDataCollator,
         DataCollatorForLanguageModeling,
         GlueDataset,
+        GlueDataTrainingArguments,
         TextDataset,
     )
 
@@ -26,7 +27,7 @@ class DataCollatorIntegrationTest(unittest.TestCase):
     def test_default_classification(self):
         MODEL_ID = "bert-base-cased-finetuned-mrpc"
         tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
-        data_args = DataTrainingArguments(
+        data_args = GlueDataTrainingArguments(
             task_name="mrpc", data_dir="./examples/tests_samples/MRPC", overwrite_cache=True
         )
         dataset = GlueDataset(data_args, tokenizer=tokenizer, evaluate=True)
@@ -37,7 +38,7 @@ class DataCollatorIntegrationTest(unittest.TestCase):
     def test_default_regression(self):
         MODEL_ID = "distilroberta-base"
         tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
-        data_args = DataTrainingArguments(
+        data_args = GlueDataTrainingArguments(
             task_name="sts-b", data_dir="./examples/tests_samples/STS-B", overwrite_cache=True
         )
         dataset = GlueDataset(data_args, tokenizer=tokenizer, evaluate=True)
@@ -89,7 +90,7 @@ class TrainerIntegrationTest(unittest.TestCase):
         MODEL_ID = "bert-base-cased-finetuned-mrpc"
         tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
         model = AutoModelForSequenceClassification.from_pretrained(MODEL_ID)
-        data_args = DataTrainingArguments(
+        data_args = GlueDataTrainingArguments(
             task_name="mrpc", data_dir="./examples/tests_samples/MRPC", overwrite_cache=True
         )
         eval_dataset = GlueDataset(data_args, tokenizer=tokenizer, evaluate=True)
